Updated Spl/it. Improved Interface.

+++ jsf-project/src/splitPackage/BillManager.java
@@ -7,6 +7,7 @@ import java.util.List;
 import java.io.Serializable;
 
 import splitPackageJDBC.JDBCSQLiteConnection;
+
 import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -19,8 +20,10 @@ import java.sql.Statement;
 public class BillManager extends ApplicationManager implements Serializable {
 	private User currentUser;
 	private Bill currentBill;
-	private List<Bill> bList;
-	private List<Bill> bOwedList;
+	private List<Bill> billsList;
+	private List<Bill> owedToYouList;
+	private String totalPay;
+	private String totalReceive; 
 	private String statusMessage;
 	private String removeID;
 	private String recipientName;
@@ -29,8 +32,8 @@ public class BillManager extends ApplicationManager implements Serializable {
 	// Constructors
 	public BillManager() {
 		currentBill = new Bill();
-		bList = new ArrayList<Bill>();
-		bOwedList = new ArrayList<Bill>();
+		billsList = new ArrayList<Bill>();
+		owedToYouList = new ArrayList<Bill>();
 		recipientList = new ArrayList<User>();
 	}
 
@@ -39,62 +42,27 @@ public class BillManager extends ApplicationManager implements Serializable {
 	}
 
 	// Getters and Setters
-	public User getCurrentUser() {
-		return currentUser;
-	}
-
-	public void setCurrentUser(User currentUser) {
-		this.currentUser = currentUser;
-	}
-
-	public Bill getCurrentBill() {
-		return currentBill;
-	}
-
-	public void setCurrentBill(Bill currentBill) {
-		this.currentBill = currentBill;
-	}
-
-	public String getStatusMessage() {
-		return statusMessage;
-	}
-
-	public void setStatusMessage(String statusMessage) {
-		this.statusMessage = statusMessage;
-	}
-
-	public String getRecipientName() {
-		return recipientName;
-	}
-	public void setRecipientName(String name) {
-		recipientName = name;
-	}
-
-
-	public void setRecipientList(String recipientName) {
-		this.recipientName = recipientName;
-	}
-
-	public String getRemoveID() {
-		return removeID;
-	}
-
-	public void setRemoveID(String removeID) {
-		this.removeID = removeID;
-	}
-
-	public void setbList(List<Bill> bList) {
-		this.bList = bList;
-	}
-	public void setbOwedList(List<Bill> bOwedList){
-		this.bOwedList = bOwedList;
-	}
-	public List<User> getRecipientList() {
-		return this.recipientList;
-	}
-	public void setRecipientList(List<User> reps) {
-		this.recipientList = reps;
-	}	
+	public User getCurrentUser() { return currentUser; }
+	public void setCurrentUser(User currentUser) { this.currentUser = currentUser; }
+	public Bill getCurrentBill() { return currentBill; }
+	public void setCurrentBill(Bill currentBill) { this.currentBill = currentBill;}
+	public String getStatusMessage() { return statusMessage; }
+	public void setStatusMessage(String statusMessage) { this.statusMessage = statusMessage; } 
+	public String getRecipientName() { return recipientName; }
+	public void setRecipientName(String name) { recipientName = name; }
+	public void setRecipientList(String recipientName) { this.recipientName = recipientName; }
+	public String getRemoveID() { return removeID; }
+	public void setRemoveID(String removeID) { this.removeID = removeID; }
+	public List<Bill> getBillsList() { return this.generateBillsList(); }
+	public List<Bill> getOwedToYouList() { return this.generateOwedToYouList(); }
+	public void setbList(List<Bill> bList) { this.billsList = bList; }
+	public void setbOwedList(List<Bill> bOwedList) { this.owedToYouList = bOwedList; }
+	public List<User> getRecipientList() { return this.recipientList; }
+	public void setRecipientList(List<User> reps) { this.recipientList = reps; }	
+	public String getTotalPay() { return this.totalPay; }
+	public void setTotalPay(String amount) { this.totalPay = amount; }
+	public String getTotalReceive() { return this.totalReceive; }
+	public void setTotalReceive(String amount) { this.totalReceive = amount; }
 	// Methods
 	/**
 	 * Creates a bill.
@@ -103,15 +71,14 @@ public class BillManager extends ApplicationManager implements Serializable {
 	 */
 	public String createBill() {
 		Connection connection = null;
+		Statement statement = null;		
 		ResultSet rs = null;
-		Statement statement = null;
 		try {
 			connection = JDBCSQLiteConnection.getConnection();
 			statement = connection.createStatement();
 			// check for duplicate bill
-			String query = "SELECT * FROM bill WHERE bill_name='"
-					+ currentBill.getBill_name().toLowerCase() + "'";
-			rs = statement.executeQuery(query);
+			String query = "";
+			rs = searchTable("bill","bill_name",currentBill.getBill_name().toLowerCase());
 			if (rs.next()) {
 				if ((currentBill.getBill_name().toLowerCase()).equals(rs
 						.getString("bill_name").toLowerCase())) {
@@ -120,6 +87,7 @@ public class BillManager extends ApplicationManager implements Serializable {
 					return "addbill";
 				}
 			}
+			rs.close();
 			// Checks for correct input format
 			if (checkValidInput(currentBill.getBill_name())) {
 				statusMessage = "Invalid Bill Name";
@@ -133,21 +101,20 @@ public class BillManager extends ApplicationManager implements Serializable {
 			
 			// Splits the bill
 			currentBill.setCost(split(recipientList.size(),
-					currentBill.getTotal()));
+			currentBill.getTotal()));
 			// Inserts a bill into table bill with values
 			// bill_id, bill_name, sender_id, recipient_id, cost, total, status,
 			// date, comment
 			query = "INSERT INTO bill VALUES(null,'"
 					+ currentBill.getBill_name() + "'," + currentUser.getID()
 					+ "," + "null," + currentBill.getCost() + ","
-					+ currentBill.getTotal() + ",'Owed'," + "null," + "null,"
+					+ currentBill.getTotal() + ",'Not Paid'," + "null," + "null,"
 					+ recipientList.size() + ")";
 			statement.executeUpdate(query);
-			
-			query = "SELECT * FROM BILL WHERE bill_name='"
-					+ currentBill.getBill_name() + "'";
-			rs = statement.executeQuery(query);
+		
+			rs=searchTable("bill","bill_name",currentBill.getBill_name());
 			currentBill.setBill_ID(rs.getInt("bill_id"));
+			rs.close();
 			
 			// Inserts a bill_recipient for each recipient
 			for (User recipient : recipientList) {
@@ -172,37 +139,26 @@ public class BillManager extends ApplicationManager implements Serializable {
 		}
 		statusMessage = "Success!";
 		currentBill = new Bill();
+		generateOwedToYouList();
+		recipientList = new ArrayList<User>();
 		return "addbill";
 	}
 
 	/**
-	 * Getter for bList connects to the database and obtains all the bills owned
-	 * by the user.
+	 * This method connects to the database and obtains all the bills that 
+	 * the user needs to pay off.
 	 * 
 	 * @return List of bills.
 	 */
-	public List<Bill> getbList() {
-		List<Bill> bList = new ArrayList<Bill>();
-		Connection connection = null;
+	public List<Bill> generateBillsList() {
+		List<Bill> billsList = new ArrayList<Bill>();
 		ResultSet rs = null;
 		ResultSet rs2 = null;
-		Statement statement = null;
-		Statement statement2 = null;
+		double totalPay = 0;
 		try {
-			connection = JDBCSQLiteConnection.getConnection();
-			// YOU NEED 2 SEPARATE STATEMENTS FOR 2 CONCURRING QUERY
-			// EXECUTIONS!!
-			statement = connection.createStatement();
-			statement2 = connection.createStatement();
-			String query = "SELECT * FROM bill_recipient WHERE recipient_id="
-					+ currentUser.getID();
-
-			rs = statement.executeQuery(query);
+			rs=searchTable("bill_recipient","recipient_id",currentUser.getID());
 			while (rs.next()) {
-				// System.out.println(rs.getInt("bill_id"));
-				String query2 = "SELECT * FROM bill WHERE bill_id="
-						+ rs.getInt("bill_id");
-				rs2 = statement2.executeQuery(query2);
+				rs2=searchTable("bill","bill_id",rs.getInt("bill_id"));
 				while (rs2.next()) {
 					Bill bill = new Bill();
 					bill.setBill_ID(rs2.getInt("bill_id"));
@@ -212,51 +168,42 @@ public class BillManager extends ApplicationManager implements Serializable {
 					bill.setCost(rs2.getDouble("cost"));
 					bill.setTotal(rs2.getDouble("total"));
 					bill.setStatus(rs2.getString("status"));
-				//	bill.setNumRecipients(rs2.getInt("num_recipients"));
-					bList.add(bill);
+					totalPay += bill.getCost();
+					billsList.add(bill);
+					rs2.close();
 				}
-
 			}
+			this.totalPay = "$" +String.format("%.2f",totalPay);
 		} catch (SQLException ex) {
 			ex.printStackTrace();
 		} finally {
-			if (connection != null) {
+			if (rs != null && rs2!=null) {
 				try {
-					statement.close();
-					connection.close();
+					rs.close();
+					rs2.close();
 				} catch (SQLException e) {
 					e.printStackTrace();
 				}
 			}
 		}
-		return bList;
+		return billsList;
 	}
 
 	/**
-	 * Generates bOwedList
+	 * This method connects to the database and obtains all the bills that 
+	 * the user is awaiting payment from.
 	 * @return List<Bill>
 	 */
-	public List<Bill> getbOwedList() {
-		List<Bill> bOwedList = new ArrayList<Bill>();
-		Connection connection = null;
+	public List<Bill> generateOwedToYouList() {
+		List<Bill> owedToYouList = new ArrayList<Bill>();
 		ResultSet rs = null;
 		ResultSet rs2 = null;
-		Statement statement = null;
-		Statement statement2 = null;
+		double totalReceive = 0.0;
 		try {
-			connection = JDBCSQLiteConnection.getConnection();
-			// YOU NEED 2 SEPARATE STATEMENTS FOR 2 CONCURRING QUERY
-			// EXECUTIONS!!
-			statement = connection.createStatement();
-			statement2 = connection.createStatement();
-			String query = "SELECT * FROM bill WHERE sender_id="
-					+ currentUser.getID();
-
-			rs = statement.executeQuery(query);
+			rs = searchTable("bill", "sender_id", currentUser.getID());
 			while (rs.next()) {
-				String query2 = "SELECT * FROM bill_recipient WHERE bill_id="
-						+ rs.getInt("bill_id");
-				rs2 = statement2.executeQuery(query2);
+				rs2 = searchTable("bill_recipient", "bill_id",
+						rs.getInt("bill_id"));
 				while (rs2.next()) {
 					Bill bill = new Bill();
 					bill.setBill_ID(rs2.getInt("bill_id"));
@@ -265,23 +212,26 @@ public class BillManager extends ApplicationManager implements Serializable {
 					bill.setCost(rs.getDouble("cost"));
 					bill.setSender_ID(rs.getInt("sender_id"));
 					bill.setTotal(rs.getDouble("total"));
-					bOwedList.add(bill);
+					totalReceive += bill.getCost();
+					owedToYouList.add(bill);
+					rs2.close();
 				}
 
 			}
+			this.totalReceive = "$" + String.format("%.2f",totalReceive);
 		} catch (SQLException ex) {
 			ex.printStackTrace();
 		} finally {
-			if (connection != null) {
+			if (rs != null && rs2!=null) {
 				try {
-					statement.close();
-					connection.close();
+					rs.close();
+					rs2.close();
 				} catch (SQLException e) {
 					e.printStackTrace();
 				}
 			}
 		}
-		return bOwedList;
+		return owedToYouList;
 	}
 
 	
@@ -293,14 +243,26 @@ public class BillManager extends ApplicationManager implements Serializable {
 	public String payBill() {
 		Connection connection = null;
 		Statement statement = null;
+		ResultSet rs = null;
 		try {
 			connection = JDBCSQLiteConnection.getConnection();
 			statement = connection.createStatement();
+			try {
+				rs = searchForRecipient(Integer.parseInt(removeID),currentUser.getID());
 			// check for duplicate bill
-			String query = "DELETE FROM bill_recipient WHERE bill_id="
-					+ removeID + " AND recipient_id=" + currentUser.getID();
-			statement.executeUpdate(query);
-			checkBills();
+			if(rs.next()) {
+				String query = "DELETE FROM bill_recipient WHERE bill_id="
+						+ removeID + " AND recipient_id=" + currentUser.getID();
+				statement.executeUpdate(query);
+				checkBills();
+			} else {
+				statusMessage = "Invalid Bill_ID. Please enter another Bill_ID.";
+				return "billsyouowe";
+			}
+			} catch(NumberFormatException e) {
+				statusMessage = "Invalid Bill_ID. Please enter another Bill_ID.";
+				return "billsyouowe";
+			}
 		} catch (SQLException e) {
 			e.printStackTrace();
 		} finally {
@@ -314,21 +276,16 @@ public class BillManager extends ApplicationManager implements Serializable {
 
 			}
 		}
-		// statusMessage = "Successfully Paid!";
+		statusMessage = "";
+		setRemoveID("");
 		return "front-page";
 	}
 	
 	public String addRecipient() {
-		Connection connection = null;
-		Statement statement = null;
 		ResultSet rs = null;
-		String query;
+		//String query;
 		try {
-			connection = JDBCSQLiteConnection.getConnection();
-			statement = connection.createStatement();
-			query = "SELECT * FROM user WHERE user_name='" + recipientName
-					+ "'";
-			rs = statement.executeQuery(query);
+			rs = searchTable("user","user_name",recipientName);
 			if (rs.next()) {
 				User temp = new User();
 				temp.setUser(rs.getString("user_name"));
@@ -343,8 +300,7 @@ public class BillManager extends ApplicationManager implements Serializable {
 				} else {
 					statusMessage = "User " + temp.getUser() + " is already a recipient.";
 					return "addbill";
-				} 
-				//rs.close();
+				} ;
 			}
 		} catch (SQLException e) {
 			e.printStackTrace(); 
@@ -363,28 +319,25 @@ public class BillManager extends ApplicationManager implements Serializable {
 	public Boolean checkBills(){
 		Boolean billAllPaid = false;
 		Connection connection = null;
-		Statement statement = null;
 		ResultSet rs = null;
 		try {
-			connection = JDBCSQLiteConnection.getConnection();
-			statement = connection.createStatement();
 			// check for duplicate bill
-			String query = "SELECT * FROM bill_recipient WHERE bill_id="
-							+removeID;
-			rs = statement.executeQuery(query);
+			connection = JDBCSQLiteConnection.getConnection();
+			rs=searchTable("bill_recipient","bill_id",Integer.parseInt(removeID));
 			if(!rs.next()){
-				Statement statement2 = connection.createStatement();
-				String query2 = "DELETE FROM bill WHERE bill_id="
+				Statement statement = connection.createStatement();
+				String query = "DELETE FROM bill WHERE bill_id="
 						+ removeID;
-				statement2.executeUpdate(query2);
+				statement.executeUpdate(query);
 				billAllPaid = true;	
+				statement.close();
 			}
+			rs.close();
 		} catch (SQLException e) {
 			e.printStackTrace();
 		} finally {
 			if (connection != null) {
 				try {
-					statement.close();
 					connection.close();
 				} catch (SQLException e) {
 					e.printStackTrace();
@@ -404,8 +357,6 @@ public class BillManager extends ApplicationManager implements Serializable {
 		}
 		return false;
 	}
-	
-
 
 	private double split(int recipientAmount, double total) {
 		double cost = total / recipientAmount;
